Event.vueのロジックを関数に抽出
